HBX-1581: Move interface 'org.hibernate.cfg.reveng.dialect.MetaDataDialect' to 'org.hibernate.tool.api.reveng.MetaDataDialect'
