+++ main/src/main/java/org/hibernate/cfg/JDBCReaderFactory.java
@@ -4,11 +4,11 @@ import java.util.Properties;
 
 import org.hibernate.cfg.reveng.JDBCReader;
 import org.hibernate.cfg.reveng.ReverseEngineeringStrategy;
-import org.hibernate.cfg.reveng.dialect.MetaDataDialect;
 import org.hibernate.engine.jdbc.connections.spi.ConnectionProvider;
 import org.hibernate.engine.jdbc.spi.JdbcServices;
 import org.hibernate.exception.spi.SQLExceptionConverter;
 import org.hibernate.service.ServiceRegistry;
+import org.hibernate.tool.api.reveng.MetaDataDialect;
 
 final public class JDBCReaderFactory {
