+++ main/src/main/java/org/hibernate/cfg/reveng/AbstractDatabaseCollector.java
@@ -4,9 +4,9 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import org.hibernate.cfg.reveng.dialect.MetaDataDialect;
 import org.hibernate.mapping.ForeignKey;
 import org.hibernate.tool.api.reveng.DatabaseCollector;
+import org.hibernate.tool.api.reveng.MetaDataDialect;
 import org.hibernate.tool.api.reveng.TableIdentifier;
 
 public abstract class AbstractDatabaseCollector implements DatabaseCollector {
