+++ main/src/main/java/org/hibernate/cfg/reveng/IndexProcessor.java
@@ -10,11 +10,11 @@ import java.util.Map.Entry;
 
 import org.hibernate.JDBCException;
 import org.hibernate.cfg.JDBCBinderException;
-import org.hibernate.cfg.reveng.dialect.MetaDataDialect;
 import org.hibernate.mapping.Column;
 import org.hibernate.mapping.Index;
 import org.hibernate.mapping.Table;
 import org.hibernate.mapping.UniqueKey;
+import org.hibernate.tool.api.reveng.MetaDataDialect;
 import org.hibernate.tool.util.TableNameQualifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
